+++ spring-cloud-zuul-ratelimit-core/src/main/java/com/marcosbarbero/cloud/autoconfigure/zuul/ratelimit/filters/RateLimitPreFilter.java
@@ -32,16 +32,14 @@ import com.marcosbarbero.cloud.autoconfigure.zuul.ratelimit.config.RateLimitKeyG
 import com.marcosbarbero.cloud.autoconfigure.zuul.ratelimit.config.RateLimitUtils;
 import com.marcosbarbero.cloud.autoconfigure.zuul.ratelimit.config.RateLimiter;
 import com.marcosbarbero.cloud.autoconfigure.zuul.ratelimit.config.properties.RateLimitProperties;
-import com.marcosbarbero.cloud.autoconfigure.zuul.ratelimit.config.properties.RateLimitProperties.Policy;
+import com.marcosbarbero.cloud.autoconfigure.zuul.ratelimit.support.RateLimitEvent;
 import com.marcosbarbero.cloud.autoconfigure.zuul.ratelimit.support.RateLimitExceededException;
 import com.netflix.zuul.context.RequestContext;
 import java.util.Map;
-import java.util.Objects;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.cloud.netflix.zuul.filters.Route;
 import org.springframework.cloud.netflix.zuul.filters.RouteLocator;
-import org.springframework.context.ApplicationEvent;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.util.UrlPathHelper;
@@ -127,26 +125,5 @@ public class RateLimitPreFilter extends AbstractRateLimitFilter {
         });
 
         return null;
-    }
-
-    public class RateLimitEvent extends ApplicationEvent {
-      private static final long serialVersionUID = 5241485625003998587L;
-
-      private final Policy policy;
-      private final String remoteAddress;
-
-      public RateLimitEvent(RateLimitPreFilter source, Policy policy, String remoteAddress) {
-        super(source);
-        this.policy = Objects.requireNonNull(policy, "Policy should not be null.");
-        this.remoteAddress = Objects.requireNonNull(remoteAddress, "RemoteAddress should not be null.");
-      }
-
-      public Policy getPolicy() {
-        return this.policy;
-      }
-
-      public String getRemoteAddress() {
-        return this.remoteAddress;
-      }
     }
 }
